iDRAC update logging and status
Added some logic to grab all of the lifecycle events.  Find out which ones are
new (JID > the InstallFromRepo event)  Track the for complete based on the array
of updates we are looking for

Currently only have iDRAC, BIOS, and lifecycle

Code is super sloppy. work in progress but it is running