Handle all types/components/everythig firmware
When the DCIM_SoftwareInstallationService.InstallFromRepository method is called, it creates an UpdateList.
This can be retrieved with GetRepoBasedUpdateList() method. Parse this for the "Target" value.
This is the value that the Lifecycle uses for it's "Name" element (it just appends "update:") in it's xml response.
Tracing all new occurances of these Target values, we can track overall status for every component that
the user wants to update. Party time.