ASM-1672 If FCOEOffload is originally enabled in target server
If FCOEOffload is originally enabled in target server  and we clone from server with iscsi offload enabled the setting is not applied.

For case where reference server configuration is required to be applied, still we were performing the export configuration of the target server which was not used.
Added a skip for the configuration export in case of reference config XML is available in the input.

In case NIC partition has FCOE enabled and the configuration XML required iSCSI offload to be enabled
then in the first attempt all the offload configuration is removed and in the second attempt the required configuration is applied.