ASM-2896 do not block for RAID background initialization
After creation of a new virtual disk, a RAID background
initialization job will start after a few minutes to write parity
data. That operation can take hours to days to complete. The new
idrac code to monitor the virtual disk operations was blocking
waiting for that operation to complete. The virtual disk should
be usable while that operation is running, so this commit changes
the polling code to ignore the background initialization job.