ASM-4431 HyperV iSCSI failure when 57800 is skipped by ASM
Skipping the import configuration operation on the FQDDs for which MAC address information is not accessible. We are skipping the MAC address information for NIC 3 and 4 for 57800. This change will help in skipping the configuration for these NICS. The attributes set for other NPAR NIC configuration is not applicable to 1GB NICs from 57800 NIC Card