ASM-6719 Import configuration operation failing for VSAN with SD card in case HBA mode is not already set
Import configuration operation failing for VSAN with SD card in case HBA mode is not already set. This loop never got tested as we were using the servers which were atleast once configured to be used as VSAN node.
This bug got introduced while addressing the code comments. After addressing the code comments this part was not tested.