ASM-6720 Config XML import operation fails to set HBA mode when Virtual Disks are present
Included "RAIDresetconfig" attribute with the RAID controller while moving the Controller from RAID to HBA mode. This will ensure all the Virtual Disks are removed before the HBA mode is configured.
In case there is any Virtual Disk on the RAID controller, then HBA mode cannot be set.

Also, then Controller is already in HBA mode then RAIDresetconfig is not allowed by iDRAC that why we are limiting the RAIDresetconfig attribute only in case we are moving controller from RAID to HBA mode.