ASM-6942 VSAN iDRAC configuration fails when RAID controller mode is not in HBA Mode
When RAID Controller is not in HBA mode, we have to set it to HBA and RAIDReset Attribute needs to be passed to ensure all the Virtual Disk and Hotspare configuration is removed.
Earlier code was trying to set the RAIDReset for all the RAID controllers which is not supported by iDRAC also initialization of the variable was not correct.