ASM-7035 VSAN deployment fails with VSAN partition on back side disk
Run import configuration twice to take care of removing backplane raid configuration. With 13G servers there is no way to initialize the disk in AHCI and ATA mode. For now we are deleting the VD when server is in Raidmode and then update the server from Raid mode to AHCI mode