ASM-7916 Support FC430 and FC630 VSAN ready nodes on local disk
Added specific check for All-Flash FC630 servers where need to perform
the OS installation for ESXi on Virtual disk managed for H330 PERC
Mini controller. We need to create a Virtual disk with RAID1
configuration on the disks managed by H330 and use the same for the OS
installation.

Currently there is no specific information is passed from UI which
indicates that we have to use the controller disk for the OS
installation. Thats why some model specific changes has been added