ASM-7589 Win2012R2 Deployment Failed with Error "Unable to create new partition or locate existing partition"
Due to presence of non-raid disk on the server, Virtual disk created for OS installation is not available at "0" disk id. To take care of this we are disabling non-raid disks from the server so that they are not available to the server.
This will also take care of the case where server has non-raid disks created from some prior deployments it is not requested in the existing deployment.