ASM-8106 SD Deployment with SD Disabled fails
Changed the logic when gathering the current BiosBootSeq and HddSeq
to actually call out to the Lifecycle controller.
The reason for this being that we were previously comparing what
we wanted to set these attributes to against the original exported config xml
from before we ran any importconfig jobs. This can sometimes result
in differences between the current config and the exported config
on subsequent import jobs.

If we thought the HddSeq or the BiosBootSeq were out of synce but
actually were not, we would mistakenly try to set this values
and create errors in the lifecycle controller