VXFM-2163 Fix server_onboard for default password warning
After an iDrac factory reset it will report a warning banner when
logging in via ssh that says "Warning: It is recommended not to use
the default user name (root) and password as it is a security
risk.". This broke the `racadm_get` parsing and thus the
`server_onboard` idrac user password setting which uses racadm.