VXFM-4166 Skip NVMe drives without controllers
NVMe drives may be connected directly to the PCIe bus without a
controller. However, they are passed in the `raid_configuration` with
a "fake" controller value of "Enclosure.Internal.0-1" in the
`virtualDisks` list:

    raid_configuration:
      virtualDisks:
      - raidLevel: nonraid
        physicalDisks:
        - Disk.Bay.21:Enclosure.Internal.0-1
        - Disk.Bay.22:Enclosure.Internal.0-1
        - Disk.Bay.23:Enclosure.Internal.0-1
        controller: Enclosure.Internal.0-1
        configuration:
          id: 94e9ea6e-80a6-4bc1-be69-541334a30e84
          raidlevel: nonraid
          comparator: minimum
          numberofdisks: 3
          disktype: requirenvme
          controllerFqdd: Enclosure.Internal.0-1
        mediaType: NVMe

In this case they need to be skipped in the same way as
HBA330-attached disks -- they require no configuration.