VXFM-4624 server onboarding should not fail while waiting for connection
This is a fix to ensure that waiting for connection does not cause
puppet apply to fail. In isolated OOB hardware management network,
exceptions cause failure before timeout has been reached.

Additionally, the provider will also wait for the /cgi-bin/discovery
endpoint to become available to ensure that the Job does not fail
on the java-layer afterward.